+++ hades/src/components/media/MediaDetail.tsx
@@ -1,6 +1,6 @@
 import type { ScraperSearchResult } from '../../api/types'
 import { MediaDetailHero } from './MediaDetailHero'
-import { LibraryDetailActions, PlayAction, WatchTogetherAction } from './LibraryDetailActions'
+import {LibraryDetailActions, PlayAction, PlayFromBeginningAction, WatchTogetherAction} from './LibraryDetailActions'
 import { LibraryAdminPanel } from './LibraryAdminPanel'
 
 interface MediaDetailProps {
@@ -18,8 +18,9 @@ export function MediaDetail({ id, content_type, discoverResult, onClose, onViewI
       content_type={content_type}
       discoverResult={discoverResult}
       onBack={onClose}
-      playButton={<>
+      playButton={media => <>
         <PlayAction id={id} content_type={content_type} discoverResult={discoverResult} />
+          <PlayFromBeginningAction id={id} content_type={content_type} discoverResult={discoverResult} media={media}/>
         <WatchTogetherAction id={id} content_type={content_type} discoverResult={discoverResult} />
       </>}
       actions={media => <LibraryDetailActions id={id} content_type={content_type} discoverResult={discoverResult} onViewInLibrary={onViewInLibrary} media={media} />}